+++ nextgisweb/gui/nodepkg/antd/index.ts
@@ -20,6 +20,7 @@ export {
     Drawer,
     Dropdown,
     Empty,
+    Flex,
     FloatButton,
     Form,
     Grid,