+++ nextgisweb/pyramid/nodepkg/fonts/index.ts
@@ -1,0 +1,4 @@
+/** @entrypoint */
+import { FontsPanel } from "./FontsPanel";
+
+export default FontsPanel;