+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changes
 - Data structure permission scope removed.
 - New resource creation dialog.
 - Support for empty geometries when importing vector layers.
+- Font management in the control panel.
 
 
 4.8.0